feat(gentegre-ai): dagilim web tarafi + duzeltmeler (474-476)
WEB:
- Ucret gridinde "+" ile acilan DAGILIM ALT SATIRI (kullanici: "1 satir cok
  sikisir"): bes kova, kapatilan/tahsil/kalan, provizyon no ve ciro dip
  toplami. Kalanlar SUNUCUDAN gelir; ekran cikarma yapmaz.
- Basvuru kartinda Sozlesme secici (tek sozlesme pasif gelir, birden fazlaysa
  zorunlu), SGK'da Devredilen Kurum, KARMA policede "SGK katkisi kullanilsin".
- `KURUM_TSS` kalkti: TSS/Karma artik alt kurum (ALT_TSS/ALT_KARMA); asama
  sirasi alt kuruma gore kurulur.
- Yeni `dagilimKovalari.ts` (kova kodlari sunucuyla birebir), belgeSatir'da
  `SatirDagilimi` tipi, `api.kurumSozlesmeleri` / `api.belgeDagit`.

DB DUZELTMELERI (dev'de gorulen uc gercek hata):
- 474 `fn_fiyat_listesi_fiyat` TABLO donduruyor, skaler sanilmisti - dagilim
  hic hesaplanamiyordu. Ayrica KDV dahil liste matraha indiriliyor ve ikinci
  iskonto carpimsal uygulaniyor.
- 475 `fn_taraf_kampanya` iki imza birden varsayilanli oldugu icin belirsizdi
  ("is not unique") - basvuru karti 500 donuyordu; sarmalayici dusuruldu.
- 476 dagilim degisince kapatma/tahsilat sayaclari da tazelenir; KAPATILMIS
  ya da TAHSIL EDILMIS satira tazeleme DOKUNMAZ - kesilmis belgeyle satir
  celismesin.

Testler: DagilimAkisTestleri (uctan uca: sozlesme -> liste -> basvuru ->
satir) 8 test; toplam API 188, web 462 gecti.

+++ Entegra/GentegreAI/api/src/Gentegre.Api/Uclar/BelgeUclari.cs
@@ -263,7 +263,7 @@ public static class BelgeUclari
                        coalesce(d.ad, '') as alt_kurum_adi,
                        s.fiyat_listesi_id, s.sgk_fiyat_listesi_id, s.sgk_kurum_id,
                        s.varsayilan_karsilama, k.tur,
-                       public.fn_dagilim_rota(k.tur, s.alt_kurum, 1) as rota
+                       public.fn_dagilim_rota(k.tur, s.alt_kurum, 1::smallint) as rota
                   from public.kurum_sozlesme s
                   join public.taraf_kurum k on k.id = s.kurum_id
                   left join public.kod_deger d
